Add demo runners emscripten
