+++ src/imgui_runner/runner_sdl.cpp
@@ -30,11 +30,6 @@ void RunnerSdl::Select_Gl_Version()
   SDL_GL_SetAttribute(SDL_GL_CONTEXT_PROFILE_MASK, SDL_GL_CONTEXT_PROFILE_CORE);
   SDL_GL_SetAttribute(SDL_GL_CONTEXT_MAJOR_VERSION, 3);
   SDL_GL_SetAttribute(SDL_GL_CONTEXT_MINOR_VERSION, 2);
-#elif defined(__EMSCRIPTEN__)
-  SDL_GL_SetAttribute(SDL_GL_CONTEXT_FLAGS, 0);
-  SDL_GL_SetAttribute(SDL_GL_CONTEXT_PROFILE_MASK, SDL_GL_CONTEXT_PROFILE_ES);
-  SDL_GL_SetAttribute(SDL_GL_CONTEXT_MAJOR_VERSION, 2);
-  SDL_GL_SetAttribute(SDL_GL_CONTEXT_MINOR_VERSION, 0);
 #else
   // GL 3.0 + GLSL 130
   // const char* glsl_version = "#version 130";
@@ -50,9 +45,6 @@ std::string RunnerSdl::GlslVersion()
 #if __APPLE__
   // GL 3.2 Core + GLSL 150
   const char* glsl_version = "#version 150";
-#elif defined(__EMSCRIPTEN__)
-  const char* glsl_version = "#version 100";
-  //const char* glsl_version = "#version 300 es";
 #else
   // GL 3.0 + GLSL 130
   const char* glsl_version = "#version 130";
