+++ src/imgui_runner_demos/example_gui.h
@@ -1,0 +1,6 @@
+#ifndef IMGUI_RUNNER_DEMOS_EXAMPLE_GUI_H
+#define IMGUI_RUNNER_DEMOS_EXAMPLE_GUI_H
+
+bool ExampleGui();
+
+#endif // IMGUI_RUNNER_DEMOS_EXAMPLE_GUI_H
