+++ src/imgui_runner_demos/imgui_runner_babylon_demo_emscripten.cpp
@@ -1,0 +1,9 @@
+#include "runner_babylon_demo.h"
+#include "imgui_runner/runner_emscripten.h"
+
+int main()
+{
+  auto runner = std::make_unique<ImGui::ImGuiRunner::RunnerEmscripten>();
+  ShowDemo(std::move(runner));
+  return 0;
+}
