+++ src/imgui_runner_demos/imgui_runner_demo_emscripten.cpp
@@ -1,0 +1,9 @@
+#include "example_gui.h"
+#include "imgui_runner/runner_emscripten.h"
+
+int main()
+{
+  ImGui::ImGuiRunner::RunnerEmscripten runner;
+  runner.ShowGui = ExampleGui;
+  runner.Run();
+}
